+++ tests/run.clj
@@ -20,6 +20,7 @@
 (require "tests/string_test")
 (require "tests/sequence_test")
 (require "tests/lazy_test")
+(require "tests/lazy_seq_ns_scope_test")
 (require "tests/macro_test")
 (require "tests/error_test")
 (require "tests/atom_test")